fixed test suite

+++ internal/item/GetItems.go
@@ -25,8 +25,8 @@ func (c *GetItemsCommand) Execute(ctx context.Context) error {
 	limit, offset := conversion.PaginationToLimitOffset(c.In.Pagination, c.service.defaultMaxPageLength, c.service.maxMaxPageLength)
 	result, err := c.service.database.GetItems(ctx, model.GetItemsParams{
 		Type:   conversion.StringToSqlNullString(c.In.Type),
-		Limit:  int32(limit),
-		Offset: int32(offset),
+		Limit:  limit,
+		Offset: offset,
 	})
 	if err != nil {
 		return err
